pyproject.toml
[tool.poetry]
name = "fecatq"
version = "0.1.0"
description = "Basic tool to extract information from the Fair-Ease Asset Catalog (a format that can be used in the terriamap catalog"
authors = ["Marc Portier <marc.portier@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pysema = {git = "https://github.com/vliz-be-opsci/py-sema.git"}
pandas = "^2.2.3"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.2"
pytest = "^8.3.5"
black = "^25.1.0"
isort = "^6.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
iddas = "fecatq.__main__:main"